webrepl: Merge websocket_helper into webrepl_cli.py.
websocket_helper.py wasn't useful as a standalone library outside of
webrepl_cli.py. Removes unused server_handshake function.
